Remove references to CIAbot.  It's dead, Jim.
per http://cia.vc :

The system is down. Sorry.

CIA.vc, the open source "version control informant", was a source code
commit notification service. It helped people collaborate more
efficiently on open source projects, by broadcasting code changes over
Internet Relay Chat.

This project first went online back in 2003, when the internet was a
very different place. Since then, IRC and Subversion have taken a back
seat to Twitter and Github. For many people, the service offered by
CIA would be just as relevant today as it was when the site first went
live; but for myself as well as many former and potential supporters
of the project, this sort of centralized fast-paced development just
isn't as compelling. Technology moves on, and I've personally moved
on.